+++ sdk/extendedlocation/azure-mgmt-extendedlocation/azure/mgmt/extendedlocation/models/_models_py3.py
@@ -10,7 +10,7 @@
 import datetime
 from typing import Any, Dict, List, Optional, TYPE_CHECKING, Union
 
-from ... import _serialization
+from .. import _serialization
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
@@ -23,7 +23,7 @@ class Resource(_serialization.Model):
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -58,10 +58,10 @@ class TrackedResource(Resource):
 
     Variables are only populated by the server, and will be ignored when sending a request.
 
-    All required parameters must be populated in order to send to Azure.
+    All required parameters must be populated in order to send to server.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -106,10 +106,10 @@ class CustomLocation(TrackedResource):  # pylint: disable=too-many-instance-attr
 
     Variables are only populated by the server, and will be ignored when sending a request.
 
-    All required parameters must be populated in order to send to Azure.
+    All required parameters must be populated in order to send to server.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -121,13 +121,13 @@ class CustomLocation(TrackedResource):  # pylint: disable=too-many-instance-attr
     :ivar location: The geo-location where the resource lives. Required.
     :vartype location: str
     :ivar identity: Identity for the resource.
-    :vartype identity: ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.Identity
+    :vartype identity: ~azure.mgmt.extendedlocation.models.Identity
     :ivar system_data: Metadata pertaining to creation and last modification of the resource.
-    :vartype system_data: ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.SystemData
+    :vartype system_data: ~azure.mgmt.extendedlocation.models.SystemData
     :ivar authentication: This is optional input that contains the authentication that should be
      used to generate the namespace.
     :vartype authentication:
-     ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CustomLocationPropertiesAuthentication
+     ~azure.mgmt.extendedlocation.models.CustomLocationPropertiesAuthentication
     :ivar cluster_extension_ids: Contains the reference to the add-on that contains charts to
      deploy CRDs and operators.
     :vartype cluster_extension_ids: list[str]
@@ -138,7 +138,7 @@ class CustomLocation(TrackedResource):  # pylint: disable=too-many-instance-attr
     :vartype host_resource_id: str
     :ivar host_type: Type of host the Custom Locations is referencing (Kubernetes, etc...).
      "Kubernetes"
-    :vartype host_type: str or ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.HostType
+    :vartype host_type: str or ~azure.mgmt.extendedlocation.models.HostType
     :ivar namespace: Kubernetes namespace that will be created on the specified cluster.
     :vartype namespace: str
     :ivar provisioning_state: Provisioning State for the Custom Location.
@@ -191,11 +191,11 @@ class CustomLocation(TrackedResource):  # pylint: disable=too-many-instance-attr
         :keyword location: The geo-location where the resource lives. Required.
         :paramtype location: str
         :keyword identity: Identity for the resource.
-        :paramtype identity: ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.Identity
+        :paramtype identity: ~azure.mgmt.extendedlocation.models.Identity
         :keyword authentication: This is optional input that contains the authentication that should be
          used to generate the namespace.
         :paramtype authentication:
-         ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CustomLocationPropertiesAuthentication
+         ~azure.mgmt.extendedlocation.models.CustomLocationPropertiesAuthentication
         :keyword cluster_extension_ids: Contains the reference to the add-on that contains charts to
          deploy CRDs and operators.
         :paramtype cluster_extension_ids: list[str]
@@ -206,7 +206,7 @@ class CustomLocation(TrackedResource):  # pylint: disable=too-many-instance-attr
         :paramtype host_resource_id: str
         :keyword host_type: Type of host the Custom Locations is referencing (Kubernetes, etc...).
          "Kubernetes"
-        :paramtype host_type: str or ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.HostType
+        :paramtype host_type: str or ~azure.mgmt.extendedlocation.models.HostType
         :keyword namespace: Kubernetes namespace that will be created on the specified cluster.
         :paramtype namespace: str
         :keyword provisioning_state: Provisioning State for the Custom Location.
@@ -224,59 +224,6 @@ class CustomLocation(TrackedResource):  # pylint: disable=too-many-instance-attr
         self.provisioning_state = provisioning_state
 
 
-class CustomLocationFindTargetResourceGroupProperties(_serialization.Model):
-    """The Find Target Resource Group operation request.
-
-    :ivar labels: Labels of the custom resource, this is a map of {key,value} pairs.
-    :vartype labels: dict[str, str]
-    """
-
-    _attribute_map = {
-        "labels": {"key": "labels", "type": "{str}"},
-    }
-
-    def __init__(self, *, labels: Optional[Dict[str, str]] = None, **kwargs: Any) -> None:
-        """
-        :keyword labels: Labels of the custom resource, this is a map of {key,value} pairs.
-        :paramtype labels: dict[str, str]
-        """
-        super().__init__(**kwargs)
-        self.labels = labels
-
-
-class CustomLocationFindTargetResourceGroupResult(_serialization.Model):
-    """The Find Target Resource Group operation response.
-
-    Variables are only populated by the server, and will be ignored when sending a request.
-
-    :ivar matched_resource_sync_rule: The matching resource sync rule is the particular resource
-     sync rule that matched the match expressions and labels and had lowest priority. This is the
-     rule responsible for mapping the target resource to the target resource group.
-    :vartype matched_resource_sync_rule: str
-    :ivar target_resource_group: The target resource group of matching resource sync rule. The
-     labels from the request will be used to find out matching resource sync rule against the
-     selector property of the resource sync rule. The one with highest priority will be returned if
-     there are multiple matching rules.
-    :vartype target_resource_group: str
-    """
-
-    _validation = {
-        "matched_resource_sync_rule": {"readonly": True},
-        "target_resource_group": {"readonly": True},
-    }
-
-    _attribute_map = {
-        "matched_resource_sync_rule": {"key": "matchedResourceSyncRule", "type": "str"},
-        "target_resource_group": {"key": "targetResourceGroup", "type": "str"},
-    }
-
-    def __init__(self, **kwargs: Any) -> None:
-        """ """
-        super().__init__(**kwargs)
-        self.matched_resource_sync_rule = None
-        self.target_resource_group = None
-
-
 class CustomLocationListResult(_serialization.Model):
     """The List Custom Locations operation response.
 
@@ -285,7 +232,7 @@ class CustomLocationListResult(_serialization.Model):
     :ivar next_link: The URL to use for getting the next set of results.
     :vartype next_link: str
     :ivar value: The list of Custom Locations.
-    :vartype value: list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CustomLocation]
+    :vartype value: list[~azure.mgmt.extendedlocation.models.CustomLocation]
     """
 
     _validation = {
@@ -361,13 +308,12 @@ class CustomLocationOperation(_serialization.Model):
 class CustomLocationOperationsList(_serialization.Model):
     """Lists of Custom Locations operations.
 
-    All required parameters must be populated in order to send to Azure.
+    All required parameters must be populated in order to send to server.
 
     :ivar next_link: Next page of operations.
     :vartype next_link: str
     :ivar value: Array of customLocationOperation. Required.
-    :vartype value:
-     list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CustomLocationOperation]
+    :vartype value: list[~azure.mgmt.extendedlocation.models.CustomLocationOperation]
     """
 
     _validation = {
@@ -386,8 +332,7 @@ class CustomLocationOperationsList(_serialization.Model):
         :keyword next_link: Next page of operations.
         :paramtype next_link: str
         :keyword value: Array of customLocationOperation. Required.
-        :paramtype value:
-         list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CustomLocationOperation]
+        :paramtype value: list[~azure.mgmt.extendedlocation.models.CustomLocationOperation]
         """
         super().__init__(**kwargs)
         self.next_link = next_link
@@ -428,7 +373,7 @@ class ProxyResource(Resource):
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -437,22 +382,6 @@ class ProxyResource(Resource):
     :vartype type: str
     """
 
-    _validation = {
-        "id": {"readonly": True},
-        "name": {"readonly": True},
-        "type": {"readonly": True},
-    }
-
-    _attribute_map = {
-        "id": {"key": "id", "type": "str"},
-        "name": {"key": "name", "type": "str"},
-        "type": {"key": "type", "type": "str"},
-    }
-
-    def __init__(self, **kwargs: Any) -> None:
-        """ """
-        super().__init__(**kwargs)
-
 
 class EnabledResourceType(ProxyResource):
     """EnabledResourceType definition.
@@ -460,7 +389,7 @@ class EnabledResourceType(ProxyResource):
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -468,14 +397,14 @@ class EnabledResourceType(ProxyResource):
      "Microsoft.Storage/storageAccounts".
     :vartype type: str
     :ivar system_data: Metadata pertaining to creation and last modification of the resource.
-    :vartype system_data: ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.SystemData
+    :vartype system_data: ~azure.mgmt.extendedlocation.models.SystemData
     :ivar cluster_extension_id: Cluster Extension ID.
     :vartype cluster_extension_id: str
     :ivar extension_type: Cluster Extension Type.
     :vartype extension_type: str
     :ivar types_metadata: Metadata of the Resource Type.
     :vartype types_metadata:
-     list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.EnabledResourceTypePropertiesTypesMetadataItem]
+     list[~azure.mgmt.extendedlocation.models.EnabledResourceTypePropertiesTypesMetadataItem]
     """
 
     _validation = {
@@ -513,7 +442,7 @@ class EnabledResourceType(ProxyResource):
         :paramtype extension_type: str
         :keyword types_metadata: Metadata of the Resource Type.
         :paramtype types_metadata:
-         list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.EnabledResourceTypePropertiesTypesMetadataItem]
+         list[~azure.mgmt.extendedlocation.models.EnabledResourceTypePropertiesTypesMetadataItem]
         """
         super().__init__(**kwargs)
         self.system_data = None
@@ -522,7 +451,7 @@ class EnabledResourceType(ProxyResource):
         self.types_metadata = types_metadata
 
 
-class EnabledResourceTypePropertiesTypesMetadataItem(_serialization.Model):
+class EnabledResourceTypePropertiesTypesMetadataItem(_serialization.Model):  # pylint: disable=name-too-long
     """Metadata of the Resource Type.
 
     :ivar api_version: Api Version of Resource Type.
@@ -569,8 +498,7 @@ class EnabledResourceTypesListResult(_serialization.Model):
     :ivar next_link: The URL to use for getting the next set of results.
     :vartype next_link: str
     :ivar value: The list of EnabledResourceTypes available for a customLocation.
-    :vartype value:
-     list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.EnabledResourceType]
+    :vartype value: list[~azure.mgmt.extendedlocation.models.EnabledResourceType]
     """
 
     _validation = {
@@ -630,10 +558,9 @@ class ErrorDetail(_serialization.Model):
     :ivar target: The error target.
     :vartype target: str
     :ivar details: The error details.
-    :vartype details: list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ErrorDetail]
+    :vartype details: list[~azure.mgmt.extendedlocation.models.ErrorDetail]
     :ivar additional_info: The error additional info.
-    :vartype additional_info:
-     list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ErrorAdditionalInfo]
+    :vartype additional_info: list[~azure.mgmt.extendedlocation.models.ErrorAdditionalInfo]
     """
 
     _validation = {
@@ -667,7 +594,7 @@ class ErrorResponse(_serialization.Model):
     operations. (This also follows the OData error response format.).
 
     :ivar error: The error object.
-    :vartype error: ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ErrorDetail
+    :vartype error: ~azure.mgmt.extendedlocation.models.ErrorDetail
     """
 
     _attribute_map = {
@@ -677,7 +604,7 @@ class ErrorResponse(_serialization.Model):
     def __init__(self, *, error: Optional["_models.ErrorDetail"] = None, **kwargs: Any) -> None:
         """
         :keyword error: The error object.
-        :paramtype error: ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ErrorDetail
+        :paramtype error: ~azure.mgmt.extendedlocation.models.ErrorDetail
         """
         super().__init__(**kwargs)
         self.error = error
@@ -693,8 +620,7 @@ class Identity(_serialization.Model):
     :ivar tenant_id: The tenant ID of resource.
     :vartype tenant_id: str
     :ivar type: The identity type. Known values are: "SystemAssigned" and "None".
-    :vartype type: str or
-     ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ResourceIdentityType
+    :vartype type: str or ~azure.mgmt.extendedlocation.models.ResourceIdentityType
     """
 
     _validation = {
@@ -711,8 +637,7 @@ class Identity(_serialization.Model):
     def __init__(self, *, type: Optional[Union[str, "_models.ResourceIdentityType"]] = None, **kwargs: Any) -> None:
         """
         :keyword type: The identity type. Known values are: "SystemAssigned" and "None".
-        :paramtype type: str or
-         ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ResourceIdentityType
+        :paramtype type: str or ~azure.mgmt.extendedlocation.models.ResourceIdentityType
         """
         super().__init__(**kwargs)
         self.principal_id = None
@@ -720,58 +645,17 @@ class Identity(_serialization.Model):
         self.type = type
 
 
-class MatchExpressionsProperties(_serialization.Model):
-    """Resource Sync Rules matchExpression property definition.
-
-    :ivar key: Key is the label key that the selector applies to.
-    :vartype key: str
-    :ivar operator: The Operator field represents a key's relationship to a set of values. Valid
-     operators are In, NotIn, Exists and DoesNotExist.
-    :vartype operator: str
-    :ivar values: The label value.
-    :vartype values: list[str]
-    """
-
-    _attribute_map = {
-        "key": {"key": "key", "type": "str"},
-        "operator": {"key": "operator", "type": "str"},
-        "values": {"key": "values", "type": "[str]"},
-    }
-
-    def __init__(
-        self,
-        *,
-        key: Optional[str] = None,
-        operator: Optional[str] = None,
-        values: Optional[List[str]] = None,
-        **kwargs: Any
-    ) -> None:
-        """
-        :keyword key: Key is the label key that the selector applies to.
-        :paramtype key: str
-        :keyword operator: The Operator field represents a key's relationship to a set of values. Valid
-         operators are In, NotIn, Exists and DoesNotExist.
-        :paramtype operator: str
-        :keyword values: The label value.
-        :paramtype values: list[str]
-        """
-        super().__init__(**kwargs)
-        self.key = key
-        self.operator = operator
-        self.values = values
-
-
 class PatchableCustomLocations(_serialization.Model):
     """The Custom Locations patchable resource definition.
 
     :ivar identity: Identity for the resource.
-    :vartype identity: ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.Identity
+    :vartype identity: ~azure.mgmt.extendedlocation.models.Identity
     :ivar tags: Resource tags.
     :vartype tags: dict[str, str]
     :ivar authentication: This is optional input that contains the authentication that should be
      used to generate the namespace.
     :vartype authentication:
-     ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CustomLocationPropertiesAuthentication
+     ~azure.mgmt.extendedlocation.models.CustomLocationPropertiesAuthentication
     :ivar cluster_extension_ids: Contains the reference to the add-on that contains charts to
      deploy CRDs and operators.
     :vartype cluster_extension_ids: list[str]
@@ -782,7 +666,7 @@ class PatchableCustomLocations(_serialization.Model):
     :vartype host_resource_id: str
     :ivar host_type: Type of host the Custom Locations is referencing (Kubernetes, etc...).
      "Kubernetes"
-    :vartype host_type: str or ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.HostType
+    :vartype host_type: str or ~azure.mgmt.extendedlocation.models.HostType
     :ivar namespace: Kubernetes namespace that will be created on the specified cluster.
     :vartype namespace: str
     :ivar provisioning_state: Provisioning State for the Custom Location.
@@ -817,13 +701,13 @@ class PatchableCustomLocations(_serialization.Model):
     ) -> None:
         """
         :keyword identity: Identity for the resource.
-        :paramtype identity: ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.Identity
+        :paramtype identity: ~azure.mgmt.extendedlocation.models.Identity
         :keyword tags: Resource tags.
         :paramtype tags: dict[str, str]
         :keyword authentication: This is optional input that contains the authentication that should be
          used to generate the namespace.
         :paramtype authentication:
-         ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CustomLocationPropertiesAuthentication
+         ~azure.mgmt.extendedlocation.models.CustomLocationPropertiesAuthentication
         :keyword cluster_extension_ids: Contains the reference to the add-on that contains charts to
          deploy CRDs and operators.
         :paramtype cluster_extension_ids: list[str]
@@ -834,7 +718,7 @@ class PatchableCustomLocations(_serialization.Model):
         :paramtype host_resource_id: str
         :keyword host_type: Type of host the Custom Locations is referencing (Kubernetes, etc...).
          "Kubernetes"
-        :paramtype host_type: str or ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.HostType
+        :paramtype host_type: str or ~azure.mgmt.extendedlocation.models.HostType
         :keyword namespace: Kubernetes namespace that will be created on the specified cluster.
         :paramtype namespace: str
         :keyword provisioning_state: Provisioning State for the Custom Location.
@@ -852,274 +736,6 @@ class PatchableCustomLocations(_serialization.Model):
         self.provisioning_state = provisioning_state
 
 
-class PatchableResourceSyncRule(_serialization.Model):
-    """The Resource Sync Rules patchable resource definition.
-
-    Variables are only populated by the server, and will be ignored when sending a request.
-
-    :ivar tags: Resource tags.
-    :vartype tags: dict[str, str]
-    :ivar priority: Priority represents a priority of the Resource Sync Rule.
-    :vartype priority: int
-    :ivar provisioning_state: Provisioning State for the Resource Sync Rule.
-    :vartype provisioning_state: str
-    :ivar selector: A label selector is composed of two parts, matchLabels and matchExpressions.
-     The first part, matchLabels is a map of {key,value} pairs. A single {key,value} in the
-     matchLabels map is equivalent to an element of matchExpressions, whose key field is 'key', the
-     operator is 'In', and the values array contains only 'value'. The second part, matchExpressions
-     is a list of resource selector requirements. Valid operators include In, NotIn, Exists, and
-     DoesNotExist. The values set must be non-empty in the case of In and NotIn. The values set must
-     be empty in the case of Exists and DoesNotExist. All of the requirements, from both matchLabels
-     and matchExpressions must all be satisfied in order to match.
-    :vartype selector:
-     ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ResourceSyncRulePropertiesSelector
-    :ivar target_resource_group: For an unmapped custom resource, its labels will be used to find
-     matching resource sync rules. If this resource sync rule is one of the matching rules with
-     highest priority, then the unmapped custom resource will be projected to the target resource
-     group associated with this resource sync rule. The user creating this resource sync rule should
-     have write permissions on the target resource group and this write permission will be validated
-     when creating the resource sync rule.
-    :vartype target_resource_group: str
-    """
-
-    _validation = {
-        "provisioning_state": {"readonly": True},
-    }
-
-    _attribute_map = {
-        "tags": {"key": "tags", "type": "{str}"},
-        "priority": {"key": "properties.priority", "type": "int"},
-        "provisioning_state": {"key": "properties.provisioningState", "type": "str"},
-        "selector": {"key": "properties.selector", "type": "ResourceSyncRulePropertiesSelector"},
-        "target_resource_group": {"key": "properties.targetResourceGroup", "type": "str"},
-    }
-
-    def __init__(
-        self,
-        *,
-        tags: Optional[Dict[str, str]] = None,
-        priority: Optional[int] = None,
-        selector: Optional["_models.ResourceSyncRulePropertiesSelector"] = None,
-        target_resource_group: Optional[str] = None,
-        **kwargs: Any
-    ) -> None:
-        """
-        :keyword tags: Resource tags.
-        :paramtype tags: dict[str, str]
-        :keyword priority: Priority represents a priority of the Resource Sync Rule.
-        :paramtype priority: int
-        :keyword selector: A label selector is composed of two parts, matchLabels and matchExpressions.
-         The first part, matchLabels is a map of {key,value} pairs. A single {key,value} in the
-         matchLabels map is equivalent to an element of matchExpressions, whose key field is 'key', the
-         operator is 'In', and the values array contains only 'value'. The second part, matchExpressions
-         is a list of resource selector requirements. Valid operators include In, NotIn, Exists, and
-         DoesNotExist. The values set must be non-empty in the case of In and NotIn. The values set must
-         be empty in the case of Exists and DoesNotExist. All of the requirements, from both matchLabels
-         and matchExpressions must all be satisfied in order to match.
-        :paramtype selector:
-         ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ResourceSyncRulePropertiesSelector
-        :keyword target_resource_group: For an unmapped custom resource, its labels will be used to
-         find matching resource sync rules. If this resource sync rule is one of the matching rules with
-         highest priority, then the unmapped custom resource will be projected to the target resource
-         group associated with this resource sync rule. The user creating this resource sync rule should
-         have write permissions on the target resource group and this write permission will be validated
-         when creating the resource sync rule.
-        :paramtype target_resource_group: str
-        """
-        super().__init__(**kwargs)
-        self.tags = tags
-        self.priority = priority
-        self.provisioning_state = None
-        self.selector = selector
-        self.target_resource_group = target_resource_group
-
-
-class ResourceSyncRule(TrackedResource):
-    """Resource Sync Rules definition.
-
-    Variables are only populated by the server, and will be ignored when sending a request.
-
-    All required parameters must be populated in order to send to Azure.
-
-    :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
-    :vartype id: str
-    :ivar name: The name of the resource.
-    :vartype name: str
-    :ivar type: The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or
-     "Microsoft.Storage/storageAccounts".
-    :vartype type: str
-    :ivar tags: Resource tags.
-    :vartype tags: dict[str, str]
-    :ivar location: The geo-location where the resource lives. Required.
-    :vartype location: str
-    :ivar system_data: Metadata pertaining to creation and last modification of the resource.
-    :vartype system_data: ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.SystemData
-    :ivar priority: Priority represents a priority of the Resource Sync Rule.
-    :vartype priority: int
-    :ivar provisioning_state: Provisioning State for the Resource Sync Rule.
-    :vartype provisioning_state: str
-    :ivar selector: A label selector is composed of two parts, matchLabels and matchExpressions.
-     The first part, matchLabels is a map of {key,value} pairs. A single {key,value} in the
-     matchLabels map is equivalent to an element of matchExpressions, whose key field is 'key', the
-     operator is 'In', and the values array contains only 'value'. The second part, matchExpressions
-     is a list of resource selector requirements. Valid operators include In, NotIn, Exists, and
-     DoesNotExist. The values set must be non-empty in the case of In and NotIn. The values set must
-     be empty in the case of Exists and DoesNotExist. All of the requirements, from both matchLabels
-     and matchExpressions must all be satisfied in order to match.
-    :vartype selector:
-     ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ResourceSyncRulePropertiesSelector
-    :ivar target_resource_group: For an unmapped custom resource, its labels will be used to find
-     matching resource sync rules. If this resource sync rule is one of the matching rules with
-     highest priority, then the unmapped custom resource will be projected to the target resource
-     group associated with this resource sync rule. The user creating this resource sync rule should
-     have write permissions on the target resource group and this write permission will be validated
-     when creating the resource sync rule.
-    :vartype target_resource_group: str
-    """
-
-    _validation = {
-        "id": {"readonly": True},
-        "name": {"readonly": True},
-        "type": {"readonly": True},
-        "location": {"required": True},
-        "system_data": {"readonly": True},
-        "provisioning_state": {"readonly": True},
-    }
-
-    _attribute_map = {
-        "id": {"key": "id", "type": "str"},
-        "name": {"key": "name", "type": "str"},
-        "type": {"key": "type", "type": "str"},
-        "tags": {"key": "tags", "type": "{str}"},
-        "location": {"key": "location", "type": "str"},
-        "system_data": {"key": "systemData", "type": "SystemData"},
-        "priority": {"key": "properties.priority", "type": "int"},
-        "provisioning_state": {"key": "properties.provisioningState", "type": "str"},
-        "selector": {"key": "properties.selector", "type": "ResourceSyncRulePropertiesSelector"},
-        "target_resource_group": {"key": "properties.targetResourceGroup", "type": "str"},
-    }
-
-    def __init__(
-        self,
-        *,
-        location: str,
-        tags: Optional[Dict[str, str]] = None,
-        priority: Optional[int] = None,
-        selector: Optional["_models.ResourceSyncRulePropertiesSelector"] = None,
-        target_resource_group: Optional[str] = None,
-        **kwargs: Any
-    ) -> None:
-        """
-        :keyword tags: Resource tags.
-        :paramtype tags: dict[str, str]
-        :keyword location: The geo-location where the resource lives. Required.
-        :paramtype location: str
-        :keyword priority: Priority represents a priority of the Resource Sync Rule.
-        :paramtype priority: int
-        :keyword selector: A label selector is composed of two parts, matchLabels and matchExpressions.
-         The first part, matchLabels is a map of {key,value} pairs. A single {key,value} in the
-         matchLabels map is equivalent to an element of matchExpressions, whose key field is 'key', the
-         operator is 'In', and the values array contains only 'value'. The second part, matchExpressions
-         is a list of resource selector requirements. Valid operators include In, NotIn, Exists, and
-         DoesNotExist. The values set must be non-empty in the case of In and NotIn. The values set must
-         be empty in the case of Exists and DoesNotExist. All of the requirements, from both matchLabels
-         and matchExpressions must all be satisfied in order to match.
-        :paramtype selector:
-         ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ResourceSyncRulePropertiesSelector
-        :keyword target_resource_group: For an unmapped custom resource, its labels will be used to
-         find matching resource sync rules. If this resource sync rule is one of the matching rules with
-         highest priority, then the unmapped custom resource will be projected to the target resource
-         group associated with this resource sync rule. The user creating this resource sync rule should
-         have write permissions on the target resource group and this write permission will be validated
-         when creating the resource sync rule.
-        :paramtype target_resource_group: str
-        """
-        super().__init__(tags=tags, location=location, **kwargs)
-        self.system_data = None
-        self.priority = priority
-        self.provisioning_state = None
-        self.selector = selector
-        self.target_resource_group = target_resource_group
-
-
-class ResourceSyncRuleListResult(_serialization.Model):
-    """The List Resource Sync Rules operation response.
-
-    Variables are only populated by the server, and will be ignored when sending a request.
-
-    :ivar next_link: The URL to use for getting the next set of results.
-    :vartype next_link: str
-    :ivar value: The list of Resource Sync Rules.
-    :vartype value: list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.ResourceSyncRule]
-    """
-
-    _validation = {
-        "next_link": {"readonly": True},
-        "value": {"readonly": True},
-    }
-
-    _attribute_map = {
-        "next_link": {"key": "nextLink", "type": "str"},
-        "value": {"key": "value", "type": "[ResourceSyncRule]"},
-    }
-
-    def __init__(self, **kwargs: Any) -> None:
-        """ """
-        super().__init__(**kwargs)
-        self.next_link = None
-        self.value = None
-
-
-class ResourceSyncRulePropertiesSelector(_serialization.Model):
-    """A label selector is composed of two parts, matchLabels and matchExpressions. The first part,
-    matchLabels is a map of {key,value} pairs. A single {key,value} in the matchLabels map is
-    equivalent to an element of matchExpressions, whose key field is 'key', the operator is 'In',
-    and the values array contains only 'value'. The second part, matchExpressions is a list of
-    resource selector requirements. Valid operators include In, NotIn, Exists, and DoesNotExist.
-    The values set must be non-empty in the case of In and NotIn. The values set must be empty in
-    the case of Exists and DoesNotExist. All of the requirements, from both matchLabels and
-    matchExpressions must all be satisfied in order to match.
-
-    :ivar match_expressions: MatchExpressions is a list of resource selector requirements. Valid
-     operators include In, NotIn, Exists, and DoesNotExist. The values set must be non-empty in the
-     case of In and NotIn. The values set must be empty in the case of Exists and DoesNotExist.
-    :vartype match_expressions:
-     list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.MatchExpressionsProperties]
-    :ivar match_labels: MatchLabels is a map of {key,value} pairs. A single {key,value} in the
-     matchLabels map is equivalent to an element of matchExpressions, whose key field is 'key', the
-     operator is 'In', and the values array contains only 'value'.
-    :vartype match_labels: dict[str, str]
-    """
-
-    _attribute_map = {
-        "match_expressions": {"key": "matchExpressions", "type": "[MatchExpressionsProperties]"},
-        "match_labels": {"key": "matchLabels", "type": "{str}"},
-    }
-
-    def __init__(
-        self,
-        *,
-        match_expressions: Optional[List["_models.MatchExpressionsProperties"]] = None,
-        match_labels: Optional[Dict[str, str]] = None,
-        **kwargs: Any
-    ) -> None:
-        """
-        :keyword match_expressions: MatchExpressions is a list of resource selector requirements. Valid
-         operators include In, NotIn, Exists, and DoesNotExist. The values set must be non-empty in the
-         case of In and NotIn. The values set must be empty in the case of Exists and DoesNotExist.
-        :paramtype match_expressions:
-         list[~azure.mgmt.extendedlocation.v2021_08_31_preview.models.MatchExpressionsProperties]
-        :keyword match_labels: MatchLabels is a map of {key,value} pairs. A single {key,value} in the
-         matchLabels map is equivalent to an element of matchExpressions, whose key field is 'key', the
-         operator is 'In', and the values array contains only 'value'.
-        :paramtype match_labels: dict[str, str]
-        """
-        super().__init__(**kwargs)
-        self.match_expressions = match_expressions
-        self.match_labels = match_labels
-
-
 class SystemData(_serialization.Model):
     """Metadata pertaining to creation and last modification of the resource.
 
@@ -1127,16 +743,14 @@ class SystemData(_serialization.Model):
     :vartype created_by: str
     :ivar created_by_type: The type of identity that created the resource. Known values are:
      "User", "Application", "ManagedIdentity", and "Key".
-    :vartype created_by_type: str or
-     ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CreatedByType
+    :vartype created_by_type: str or ~azure.mgmt.extendedlocation.models.CreatedByType
     :ivar created_at: The timestamp of resource creation (UTC).
     :vartype created_at: ~datetime.datetime
     :ivar last_modified_by: The identity that last modified the resource.
     :vartype last_modified_by: str
     :ivar last_modified_by_type: The type of identity that last modified the resource. Known values
      are: "User", "Application", "ManagedIdentity", and "Key".
-    :vartype last_modified_by_type: str or
-     ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CreatedByType
+    :vartype last_modified_by_type: str or ~azure.mgmt.extendedlocation.models.CreatedByType
     :ivar last_modified_at: The timestamp of resource last modification (UTC).
     :vartype last_modified_at: ~datetime.datetime
     """
@@ -1166,16 +780,14 @@ class SystemData(_serialization.Model):
         :paramtype created_by: str
         :keyword created_by_type: The type of identity that created the resource. Known values are:
          "User", "Application", "ManagedIdentity", and "Key".
-        :paramtype created_by_type: str or
-         ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CreatedByType
+        :paramtype created_by_type: str or ~azure.mgmt.extendedlocation.models.CreatedByType
         :keyword created_at: The timestamp of resource creation (UTC).
         :paramtype created_at: ~datetime.datetime
         :keyword last_modified_by: The identity that last modified the resource.
         :paramtype last_modified_by: str
         :keyword last_modified_by_type: The type of identity that last modified the resource. Known
          values are: "User", "Application", "ManagedIdentity", and "Key".
-        :paramtype last_modified_by_type: str or
-         ~azure.mgmt.extendedlocation.v2021_08_31_preview.models.CreatedByType
+        :paramtype last_modified_by_type: str or ~azure.mgmt.extendedlocation.models.CreatedByType
         :keyword last_modified_at: The timestamp of resource last modification (UTC).
         :paramtype last_modified_at: ~datetime.datetime
         """